refactor(e2e): move schema-mode config to schema-mode-db.ts
Move configureSchemaMode() and the default DB user/password constants
from runtime-deploy.ts to schema-mode-db.ts where the rest of the
schema-mode database utilities live. runtime-deploy.ts imports and
calls the function — it no longer owns any schema-mode-specific logic.
